+++ src/components/organisms/AssetContent/MetaFull.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, ReactElement } from 'react'
+import React, { ReactElement } from 'react'
 import Time from '../../atoms/Time'
 import MetaItem from './MetaItem'
 import styles from './MetaFull.module.css'
@@ -7,23 +7,6 @@ import { useAsset } from '../../../providers/Asset'
 
 export default function MetaFull(): ReactElement {
   const { ddo, metadata, isInPurgatory } = useAsset()
-  const [assetTimeout, setAssetTimeout] = useState('')
-
-  const renderedTimeout = (timeout: number) => {
-    switch (timeout) {
-      case 0:
-        return 'Forever'
-      case 3600:
-        return '1 hour'
-      default:
-        return ''
-    }
-  }
-
-  useEffect(() => {
-    const { timeout } = ddo.findServiceByType('access').attributes.main
-    setAssetTimeout(renderedTimeout(timeout))
-  }, [ddo])
 
   return (
     <div className={styles.metaFull}>
@@ -41,7 +24,6 @@ export default function MetaFull(): ReactElement {
 
       {/* TODO: remove those 2 date items here when EditHistory component is ready */}
       <MetaItem title="Published" content={<Time date={ddo?.created} />} />
-      <MetaItem title="Availability" content={<p>{assetTimeout}</p>} />
       {ddo?.created !== ddo?.updated && (
         <MetaItem title="Updated" content={<Time date={ddo?.updated} />} />
       )}
